refactor: Small changes based on review comments.
